pyproject.toml
[project]
name = "project-database-management-iconnet"
version = "0.1.0"
description = "end to end project database management iconnet"
authors = [
    {name = "rizkyyanuark",email = "rizkyyanuarkristianto@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "streamlit (==1.41.1)",
    "sqlalchemy (==1.4.54)",
    "pydantic (==2.10.6)",
    "psycopg2-binary (==2.9.10)",
    "plotly (==5.14.1)",
    "firebase-admin (==6.6.0)",
    "pyyaml (==6.0.2)",
    "python-dotenv (==1.0.1)",
    "pyprojroot (==0.3.0)",
    "gspread (==6.2.0)",
    "oauth2client (==4.1.3)",
    "streamlit-option-menu (==0.3.2)",
    "folium (==0.14.0)",
    "streamlit-folium (==0.15.0)",
    "geopandas (==1.0.1)",
    "langchain-core (>=0.3.54)",
    "langchain-google-genai (>=2.1.2)",
    "langchain-community (==0.3.20)",
    "langgraph (==0.3.34)",
    "langchain (==0.3.21)",
    "supabase (==2.15.0)",
    "streamlit-cookies-manager (==0.2.0)",
    "ipython (==9.0.1)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
